Cleaned up mksdir.sh
You can now pass a full name and it will use the last word in the name to find
the corresponding email address and student number. This works because all
student.ru.nl email addresses are required to end in the last name. If it
doesn't work just enter something that does match and change the name manually.